+++ packages/components/atoms/f-button/test-utils/component-objects/f-button.component.js
@@ -8,6 +8,21 @@ module.exports = class Buttons extends Page {
     open (url) {
         super.open(url);
     }
+    /**
+     * @description
+     * Sets the data for the button component.
+     * 
+     * @param {Object} button
+     * @param {String} button.type
+     * @param {String} button.size
+     */
+    // open (button) {
+    //     const type = `&knob-Button%20Type=${button.type}`;
+    //     const url = button.type === 'link' ? `&knob-href=link` : '';
+    //     const size = `&knob-Button%20Size=${button.size}`;
+
+    //     browser.url(`/iframe.html?id=components-atoms-f-button--button-component${type}${url}${size}`)
+    // }
 
     waitForActionComponent () {
         super.waitForComponent(this.actionComponent);