+++ packages/components/atoms/f-button/CHANGELOG.md
@@ -4,6 +4,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+Latest – to be added to the next release
+------------------------------
+
+### Added
+- Percy visual regression tests
+
+
 v1.3.0
 ------------------------------
 *May 25, 2021*